Apply pin-term only if the pin is of permanent nature
(= when the pinner isn't under attack)

During bench run I discovered that a greater part of pin-positions does
assign the increased kingDanger inappropriately as the pinner is under
attack:
-often the pinned-piece can safely capture the pinner. Most occuring
cases are slider-pieces pinning a piece of the same type.
-sometimes the pin by itself is very harmless, i.e:
 
 +---+---+---+---+---+---+---+---+
 | r | n | b | k |   | b | n | r |
 +---+---+---+---+---+---+---+---+
 | p | p | N |   | p | p | p | p |
 +---+---+---+---+---+---+---+---+
 |   |   |   |   | q |   |   |   |
 +---+---+---+---+---+---+---+---+
 |   |   |   |   |   |   |   |   |
 +---+---+---+---+---+---+---+---+
 |   |   |   |   |   |   |   |   |
 +---+---+---+---+---+---+---+---+
 |   |   |   |   |   |   |   |   |
 +---+---+---+---+---+---+---+---+
 | P | P |   | P | P | P | P | P |
 +---+---+---+---+---+---+---+---+
 | R |   | B | Q | K | B | N | R |
 +---+---+---+---+---+---+---+---+

Pawn on e2 is pinned by the black queen, but she is en prise and
e2-pawns mobility is not limited concretely.

bench: 5646084